pyproject.toml
[tool.poetry]
name = "nonebot-plugin-looklike"
version = "0.0.1"
description = "你看我像人吗?"
authors = ["月ヶ瀬"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "nonebot_plugin_looklike" },
]
homepage = "https://github.com/tkgs0/nonebot-plugin-looklike"
repository = "https://github.com/tkgs0/nonebot-plugin-looklike"
keywords = ["nonebot", "nonebot2", "nonebot_plugin", "looklike"]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.1"

[tool.poetry.dev-dependencies]
python = "^3.11"
nonebot2 = "^2.3.1"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
implicit_reexport = true
pretty = true
show_error_codes = true
strict = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

